refactor: split client preference dtos
